[feat]: added e2e test
