+++ apps/src/lib/node/ledger/shell/finalize_block.rs
@@ -63,7 +63,6 @@ where
 
         if new_epoch {
             for id in self.proposal_data.clone() {
-                println!("Executing: {}", id);
                 let proposal_funds_key = gov_storage::get_funds_key(id);
                 let (proposal_funds_bytes, _) = self
                     .storage
@@ -81,7 +80,6 @@ where
                 if let Ok(tally_result) = tally_result {
                     match tally_result {
                         (TallyResult::Passed, Some(code)) => {
-                            println!("tally okay + code ");
                             let tx = Tx::new(code, None);
                             let tx_type =
                                 TxType::Decrypted(DecryptedTx::Decrypted(tx));
@@ -134,10 +132,6 @@ where
                                         } else {
                                             continue;
                                         };
-                                        println!(
-                                            "transfering funds...: {}",
-                                            id
-                                        );
                                         self.storage.transfer(
                                             &m1t(),
                                             funds,
@@ -528,7 +522,7 @@ where
 
         // Update evidence parameters
         let (epoch_duration, _gas) =
-            parameters::parameters::read_epoch_parameter(&self.storage)
+            parameters::read_epoch_parameter(&self.storage)
                 .expect("Couldn't read epoch duration parameters");
         let pos_params = self.storage.read_pos_params();
         let evidence_params =
